Edulink/requirements.txt
# Core Django Dependencies
Django==5.2.3
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
django-environ==0.11.2
django-filter==23.3

# Testing Dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0
coverage==7.3.2
model-bakery==1.17.0

# Security & Validation
django-ratelimit==4.1.0
Pillow==10.1.0
django-extensions==3.2.3
cryptography==41.0.7
django-axes==6.1.1
bandit==1.7.5
safety==2.3.5

# Development Tools
mypy==1.7.1
django-stubs==4.2.7
djangorestframework-stubs==3.14.5
types-requests==2.31.0.10
ruff==0.1.6
black==23.11.0
isort==5.12.0
flake8==6.1.0
pre-commit==3.6.0
ipython==8.17.2
ipdb==0.13.13

# Performance & Monitoring
django-debug-toolbar==4.2.0
django-silk==5.0.4
redis==5.0.1
django-redis==5.4.0
hiredis==2.2.3
sentry-sdk==1.38.0
structlog==23.2.0
django-prometheus==2.3.1

# Email & Notifications
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-anymail==10.2
django-notifications-hq==1.8.3

# API Documentation
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
drf-yasg==1.21.7

# File Handling
django-storages==1.14.2
boto3==1.34.0
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0

# Authentication & Authorization
django-allauth==0.57.0
django-oauth-toolkit==1.7.1

# Database & ORM
dj-database-url==2.1.0
django-cachalot==2.6.1

# Forms & UI
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0

# Admin Enhancements
django-admin-interface==0.26.1
django-import-export==3.3.1

# Deployment & Production
gunicorn==21.2.0
whitenoise==6.6.0
gevent==23.9.1
django-health-check==3.17.0

# Data Processing
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2

# HTTP & API Clients
httpx==0.25.2
requests==2.31.0
urllib3==2.1.0

# Utilities
pytz==2023.3
PyJWT==2.8.0
python-dateutil==2.8.2
slugify==0.0.1
django-phonenumber-field==7.2.0
phonumbers==8.13.26
click==8.1.7
PyYAML==6.0.1

# Validation & Serialization
cerberus==1.3.5
jsonschema==4.20.0
marshmallow==3.20.1

# Search (Optional)
# django-haystack==3.2.1
# whoosh==2.7.4
# elasticsearch==8.11.0

# WebSocket Support (Optional)
# channels==4.0.0
# channels-redis==4.1.0
# daphne==4.0.0

# Machine Learning (Optional)
# scikit-learn==1.3.2
# tensorflow==2.15.0

# Additional Security (Optional)
# django-csp==3.7
# django-permissions-policy==4.17.0

# Audit & History (Optional)
# django-auditlog==2.2.2
# django-simple-history==3.4.0